% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_featured_packages.R
\name{featured_packages}
\alias{featured_packages}
\title{featured_packages}
\usage{
featured_packages(
  input,
  output,
  session,
  d_pkg_releases,
  d_pkg_dependencies,
  d_pkg_details
)
}
\arguments{
\item{d_pkg_details}{}
}
\description{
featured_packages
}
