% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_network.R
\name{graph_network}
\alias{graph_network}
\title{graph_network}
\usage{
graph_network(
  input,
  output,
  session,
  d_pkg_dependencies,
  d_pkg_details,
  d_pkg_releases
)
}
\arguments{
\item{d_pkg_releases}{}
}
\description{
graph_network
}
