% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pretty_value_box.R
\name{pretty_value_box}
\alias{pretty_value_box}
\title{pretty_value_box}
\usage{
pretty_value_box(
  input,
  output,
  session,
  title = "-",
  value = 0,
  last_days = "-",
  date_range = "-",
  inline_bar = 0,
  trend = 0,
  trend_icon = "",
  trend_label = "-"
)
}
\arguments{
\item{trend_label}{}
}
\description{
pretty_value_box
}
