% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pretty_value_box.R
\name{pretty_value_box_ui}
\alias{pretty_value_box_ui}
\title{pretty_value_box_ui}
\usage{
pretty_value_box_ui(
  id,
  background_color = "#e2a267",
  icon_name = "question-circle"
)
}
\arguments{
\item{icon_name}{}
}
\description{
pretty_value_box_ui
}
